backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
neo4j
pinecone-client
openai
anthropic
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
redis
boto3
pytest
httpx
slack-sdk
google-api-python-client
google-auth-oauthlib
cryptography
